# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_api.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_api.aiChatRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  internal_api.aiChatResponse:
    properties:
      response:
        type: string
    type: object
  internal_api.aiProfileItemDTO:
    properties:
      confidence:
        type: number
      field_name:
        type: string
      source:
        type: string
      value:
        type: string
    type: object
  internal_api.aiProfileResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_api.aiProfileItemDTO'
        type: array
    type: object
  internal_api.aiSummaryResponse:
    properties:
      conclusion:
        type: string
      key_points:
        type: string
      message_range:
        type: string
      topic:
        type: string
      user_intent:
        type: string
    type: object
...
  internal_api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api.loginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
    type: object
...
  internal_api.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_api.refreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
    type: object
  internal_api.registerRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_api.registerResponse:
    properties:
      user_id:
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: GoIM 即时通讯系统 API — 支持私聊、群聊、朋友圈动态、AI 助手。
  title: GoIM API
  version: "1.0"
paths:
  /ai/chat:
    post:
      consumes:
      - application/json
      description: 向AI发送消息并获取非流式回复
      parameters:
      - description: 聊天消息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.aiChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 对话成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.aiChatResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: AI对话
      tags:
      - AI
  /ai/chat/stream:
    post:
      consumes:
      - application/json
      description: 以SSE流式方式与AI助手对话，实时返回生成内容
      parameters:
      - description: 聊天消息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.aiChatRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE事件流
          schema:
            type: string
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: AI流式聊天
      tags:
      - AI
  /ai/profile:
    get:
      description: 获取当前用户的AI画像条目（第二层记忆）
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.aiProfileResponse'
              type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 获取AI画像
      tags:
      - AI
  /ai/summary/{convID}:
    post:
      description: 为指定会话生成AI摘要
      parameters:
      - description: 会话ID
        in: path
        name: convID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.aiSummaryResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      security:
      - BearerAuth: []
      summary: 生成会话摘要
      tags:
      - AI
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，返回访问令牌和刷新令牌
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.loginResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.refreshResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "401":
          description: 令牌无效或已过期
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      summary: 刷新令牌
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码注册新用户
      parameters:
      - description: 注册信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.registerResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/internal_api.ApiResponse'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    description: 在 Authorization 头中携带 JWT 访问令牌，格式：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"